[FIX] #36: Handle unsigned integer underflow in spread calculation
Fixed spread calculation for crossed books (bid > ask).

Problem:
- Price is uint32_t (unsigned)
- When best_ask < best_bid (crossed book):
  spread = best_ask - best_bid = underflow
  Example: 10010 - 10020 = 4294967286 (not -10)
- test_crossed_book() failed in CI

Solution:
- Cast both prices to int64_t before subtraction
- Now correctly handles negative spreads
- spread = static_cast<double>(static_cast<int64_t>(best_ask) - static_cast<int64_t>(best_bid))

Tests: 56/56 pass (including crossed book test)

+++ include/metrics/order_book_metrics.hpp
@@ -81,7 +81,9 @@ private:
 
         // Spread
         if (snapshot.best_bid != INVALID_PRICE && snapshot.best_ask != INVALID_PRICE) {
-            metrics_.spread = static_cast<double>(snapshot.best_ask - snapshot.best_bid);
+            // Cast to signed before subtraction to handle crossed book (negative spread)
+            metrics_.spread =
+                static_cast<double>(static_cast<int64_t>(snapshot.best_ask) - static_cast<int64_t>(snapshot.best_bid));
             metrics_.mid_price =
                 (static_cast<double>(snapshot.best_bid) + static_cast<double>(snapshot.best_ask)) / 2.0;
